copying files from server generate_python_sdk.sh into here
