+++ app/desktop/studio_server/api_client/kiln_ai_server_client/models/__init__.py
@@ -6,7 +6,6 @@ from .body_start_gepa_job_v1_jobs_gepa_job_start_post_token_budget import (
     BodyStartGepaJobV1JobsGepaJobStartPostTokenBudget,
 )
 from .body_start_sample_job_v1_jobs_sample_job_start_post import BodyStartSampleJobV1JobsSampleJobStartPost
-from .check_model_supported_response import CheckModelSupportedResponse
 from .clarify_spec_input import ClarifySpecInput
 from .clarify_spec_output import ClarifySpecOutput
 from .examples_for_feedback_item import ExamplesForFeedbackItem
@@ -19,10 +18,10 @@ from .gepa_job_result_response import GEPAJobResultResponse
 from .health_health_get_response_health_health_get import HealthHealthGetResponseHealthHealthGet
 from .http_validation_error import HTTPValidationError
 from .job_start_response import JobStartResponse
-from .judge_info import JudgeInfo
 from .job_status import JobStatus
 from .job_status_response import JobStatusResponse
 from .job_type import JobType
+from .judge_info import JudgeInfo
 from .model_provider_name import ModelProviderName
 from .new_proposed_spec_edits import NewProposedSpecEdits
 from .output_file_info import OutputFileInfo
@@ -43,7 +42,6 @@ __all__ = (
     "BodyStartGepaJobV1JobsGepaJobStartPost",
     "BodyStartGepaJobV1JobsGepaJobStartPostTokenBudget",
     "BodyStartSampleJobV1JobsSampleJobStartPost",
-    "CheckModelSupportedResponse",
     "ClarifySpecInput",
     "ClarifySpecOutput",
     "ExamplesForFeedbackItem",
@@ -57,9 +55,9 @@ __all__ = (
     "HTTPValidationError",
     "JobStartResponse",
     "JobStatus",
-    "JudgeInfo",
     "JobStatusResponse",
     "JobType",
+    "JudgeInfo",
     "ModelProviderName",
     "NewProposedSpecEdits",
     "OutputFileInfo",
