--- app/desktop/studio_server/api_client/kiln_ai_server_client/models/check_model_supported_response.py
@@ -1,62 +1,0 @@
-from __future__ import annotations
-
-from collections.abc import Mapping
-from typing import Any, TypeVar
-
-from attrs import define as _attrs_define
-from attrs import field as _attrs_field
-
-T = TypeVar("T", bound="CheckModelSupportedResponse")
-
-
-@_attrs_define
-class CheckModelSupportedResponse:
-    """Response model for check_model_supported endpoint.
-
-    Attributes:
-        is_model_supported (bool):
-    """
-
-    is_model_supported: bool
-    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
-
-    def to_dict(self) -> dict[str, Any]:
-        is_model_supported = self.is_model_supported
-
-        field_dict: dict[str, Any] = {}
-        field_dict.update(self.additional_properties)
-        field_dict.update(
-            {
-                "is_model_supported": is_model_supported,
-            }
-        )
-
-        return field_dict
-
-    @classmethod
-    def from_dict(cls: type[T], src_dict: Mapping[str, Any]) -> T:
-        d = dict(src_dict)
-        is_model_supported = d.pop("is_model_supported")
-
-        check_model_supported_response = cls(
-            is_model_supported=is_model_supported,
-        )
-
-        check_model_supported_response.additional_properties = d
-        return check_model_supported_response
-
-    @property
-    def additional_keys(self) -> list[str]:
-        return list(self.additional_properties.keys())
-
-    def __getitem__(self, key: str) -> Any:
-        return self.additional_properties[key]
-
-    def __setitem__(self, key: str, value: Any) -> None:
-        self.additional_properties[key] = value
-
-    def __delitem__(self, key: str) -> None:
-        del self.additional_properties[key]
-
-    def __contains__(self, key: str) -> bool:
-        return key in self.additional_properties
