+++ app/desktop/studio_server/api_client/kiln_ai_server_client/models/examples_with_feedback_item.py
@@ -17,7 +17,7 @@ class ExamplesWithFeedbackItem:
         user_agrees_with_judge (bool): Whether the user agrees with the judge's assessment
         input_ (str):
         output (str):
-        fails_specification (bool): Whether the output fails the specification
+        fails_specification (bool): Judge's verdict - whether the output fails the Target Specification
         user_feedback (None | str | Unset): Optional text feedback from the user
     """
