+++ app/desktop/studio_server/api_client/kiln_ai_server_client/models/judge_info.py
@@ -4,6 +4,7 @@ from collections.abc import Mapping
 from typing import Any, TypeVar
 
 from attrs import define as _attrs_define
+from attrs import field as _attrs_field
 
 from ..models.model_provider_name import ModelProviderName
 
@@ -22,6 +23,7 @@ class JudgeInfo:
     model_id: str
     model_provider: ModelProviderName
     judge_prompt: str
+    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
     def to_dict(self) -> dict[str, Any]:
         model_id = self.model_id
@@ -31,7 +33,7 @@ class JudgeInfo:
         judge_prompt = self.judge_prompt
 
         field_dict: dict[str, Any] = {}
-
+        field_dict.update(self.additional_properties)
         field_dict.update(
             {
                 "model_id": model_id,
@@ -57,5 +59,21 @@ class JudgeInfo:
             judge_prompt=judge_prompt,
         )
 
+        judge_info.additional_properties = d
         return judge_info
 
+    @property
+    def additional_keys(self) -> list[str]:
+        return list(self.additional_properties.keys())
+
+    def __getitem__(self, key: str) -> Any:
+        return self.additional_properties[key]
+
+    def __setitem__(self, key: str, value: Any) -> None:
+        self.additional_properties[key] = value
+
+    def __delitem__(self, key: str) -> None:
+        del self.additional_properties[key]
+
+    def __contains__(self, key: str) -> bool:
+        return key in self.additional_properties
