+++ app/desktop/studio_server/api_client/kiln_ai_server_client/models/refine_spec_input.py
@@ -19,29 +19,17 @@ T = TypeVar("T", bound="RefineSpecInput")
 class RefineSpecInput:
     """
     Attributes:
-        task_prompt_with_few_shot (str):
-        task_input_schema (str):
-        task_output_schema (str):
         task_info (TaskInfo):
         spec (Spec):
         examples_with_feedback (list[ExamplesWithFeedbackItem]):
     """
 
-    task_prompt_with_few_shot: str
-    task_input_schema: str
-    task_output_schema: str
     task_info: TaskInfo
     spec: Spec
     examples_with_feedback: list[ExamplesWithFeedbackItem]
     additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
 
     def to_dict(self) -> dict[str, Any]:
-        task_prompt_with_few_shot = self.task_prompt_with_few_shot
-
-        task_input_schema = self.task_input_schema
-
-        task_output_schema = self.task_output_schema
-
         task_info = self.task_info.to_dict()
 
         spec = self.spec.to_dict()
@@ -55,9 +43,6 @@ class RefineSpecInput:
         field_dict.update(self.additional_properties)
         field_dict.update(
             {
-                "task_prompt_with_few_shot": task_prompt_with_few_shot,
-                "task_input_schema": task_input_schema,
-                "task_output_schema": task_output_schema,
                 "task_info": task_info,
                 "spec": spec,
                 "examples_with_feedback": examples_with_feedback,
@@ -73,12 +58,6 @@ class RefineSpecInput:
         from ..models.task_info import TaskInfo
 
         d = dict(src_dict)
-        task_prompt_with_few_shot = d.pop("task_prompt_with_few_shot")
-
-        task_input_schema = d.pop("task_input_schema")
-
-        task_output_schema = d.pop("task_output_schema")
-
         task_info = TaskInfo.from_dict(d.pop("task_info"))
 
         spec = Spec.from_dict(d.pop("spec"))
@@ -91,9 +70,6 @@ class RefineSpecInput:
             examples_with_feedback.append(examples_with_feedback_item)
 
         refine_spec_input = cls(
-            task_prompt_with_few_shot=task_prompt_with_few_shot,
-            task_input_schema=task_input_schema,
-            task_output_schema=task_output_schema,
             task_info=task_info,
             spec=spec,
             examples_with_feedback=examples_with_feedback,
